+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## BC breaks
 - Remove FlexibleEntityBundle
 - Remove CategoryWriter and use the generic doctrine writer instead
+- Remove entity argument from FiltersConfigurator constructor
 
 # 1.1.0 - "Rabbit Punch" (2014-04-16)